todo_project/requirements.txt
Django==2.2.9
psycopg2
djangorestframework
django-cors-headers